refactor(httpstub): rename stubs package to httpstub, New constructor
httpstub.New(...) reads as self-describing as http.NewServeMux —
the package name provides the protocol context so no prefix is
needed on the constructor or the type.
