+++ example_test.go
@@ -10,11 +10,11 @@ import (
 	"testing"
 
 	"github.com/kyuff/qa"
-	"github.com/kyuff/qa/stubs"
+	"github.com/kyuff/qa/httpstub"
 )
 
 // Fixed address so the app can always reach the stub, regardless of mode.
-var paymentStub = stubs.NewHTTP("payments", stubs.WithAddr("localhost:19001"))
+var paymentStub = httpstub.New("payments", httpstub.WithAddr("localhost:19001"))
 
 func TestMain(m *testing.M) {
 	// paymentStub.URL is set at construction (WithAddr), so it is ready here.
@@ -49,7 +49,7 @@ func (g *orderGiven) PaymentServiceAcceptsCharge() *orderGiven {
 	g.ctx.Run("payment service accepts charge", func(t *testing.T) {
 		paymentStub.
 			On("POST", "/charge").
-			WithBody(stubs.Contains(g.ctx.Data.OrderID)).
+			WithBody(httpstub.Contains(g.ctx.Data.OrderID)).
 			Return(200, `{"ok":true}`)
 	})
 	return g
@@ -74,7 +74,7 @@ func (th *orderThen) PaymentWasCharged() *orderThen {
 	th.ctx.Run("payment was charged", func(t *testing.T) {
 		calls := paymentStub.
 			Calls("POST", "/charge").
-			WithBody(stubs.Contains(th.ctx.Data.OrderID))
+			WithBody(httpstub.Contains(th.ctx.Data.OrderID))
 		if len(calls) == 0 {
 			t.Errorf("expected payment service to be called with order %s", th.ctx.Data.OrderID)
 		}
