+++ httpstub/matcher.go
@@ -1,4 +1,4 @@
-package stubs
+package httpstub
 
 import "bytes"
